requirements.txt
python-telegram-bot>=21.7
httpx>=0.27.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
